Nice segmentation API. Better README.

+++ Scripts/script_segment_vessels.m
@@ -1,8 +1,14 @@
 
-%script_segment_vessels
+% SCRIPT_SEGMENT_VESSELS
+% -------------------------------------------------------------------------
+% This code is used for segmenting blood vessels in fundus images. It is
+% based on our TBME paper (Orlando et al., 2016: A Discriminatively Trained
+% Fully Connected Conditional Random Field Model for Blood Vessel
+% Segmentation in Fundus Images). You have to edit config_segment_vessels
+% before running this code.
+% -------------------------------------------------------------------------
 
 config_segment_vessels
-
 
 % for each data set
 for d = 1 : length(dataset_names)
@@ -20,28 +26,31 @@ for d = 1 : length(dataset_names)
     current_output_data_folder = fullfile(output_segmentations_folder, current_dataset, 'segmentations');
     
     % copy all images to a new folder
-    fprintf('Copy all images\n');
-    standardized_size_dataset_folder = fullfile(current_image_data_folder, '_aux');
-    mkdir(standardized_size_dataset_folder);
-    copyfile(current_image_data_folder, standardized_size_dataset_folder, 'f');
+    if exist(fullfile(current_image_data_folder, '_aux'), 'dir')==0
+        fprintf('Copy all images\n');
+        standardized_size_dataset_folder = fullfile(current_image_data_folder, '_aux');
+        mkdir(standardized_size_dataset_folder);
+        copyfile(current_image_data_folder, standardized_size_dataset_folder, 'f');
     
-    % retrieve minimum image size
-    fprintf('Retrieving smallest image size\n');
-    standardized_size_images_folder = fullfile(standardized_size_dataset_folder, 'images');
-    standardized_size_masks_folder = fullfile(standardized_size_dataset_folder, 'masks');
-    [min_x, image_names_per_size] = getMinimumImageSize(standardized_size_images_folder);
-    fprintf('-- Smallest size is %i\n', min_x);
+        % retrieve minimum image size
+        fprintf('Retrieving smallest image size\n');
+        standardized_size_images_folder = fullfile(standardized_size_dataset_folder, 'images');
+        standardized_size_masks_folder = fullfile(standardized_size_dataset_folder, 'masks');
+        [min_x, image_names_per_size] = getMinimumImageSize(standardized_size_images_folder);
+        fprintf('-- Smallest size is %i\n', min_x);
     
-    % standardize image sizes
-    fprintf('Standardizing images\n');
-    standardizeDatasetSize(standardized_size_images_folder, standardized_size_masks_folder, min_x);
+        % standardize image sizes
+        fprintf('Standardizing images\n');
+        standardizeDatasetSize(standardized_size_images_folder, standardized_size_masks_folder, min_x);
+
+    end
     
     % Segment!!
     fprintf('Segmenting images\n');
     datasets_names = {fullfile(current_dataset, '_aux')};   
     scale_to_downsample = current_scale_factor;
-    rootDatasets = 'C:\_diabetic_retinopathy';
-    rootResults = 'C:\_diabetic_retinopathy\segmentations';
+    rootDatasets = image_folder;
+    rootResults = fullfile(output_segmentations_folder, dataset_names{d}, 'segmentations');
     script_evaluate_existing_model;
     
     % Copy results to a new folder
